Only use limited options from external config file
Most of the options are saved to an external file (keropi/config), But majority of the options are either only set to defaults, or is not used or relevant for libretro. This PR reduces the reliance to this external config file by only saving what is commonly should be manually set is some cases, aka StartDir="" for example.